Extract task graph status projection
