+++ loopx/projections/__init__.py
@@ -1,0 +1,1 @@
+"""Read-model projection helpers for LoopX status surfaces."""
